Move consumer parameters into separate file.
