+++ ocean_lib/services/service.py
@@ -9,13 +9,12 @@
 import copy
 import logging
 import re
-from distutils.util import strtobool
 from typing import Any, Dict, List, Optional
 
-from enforce_typing import enforce_types
 from web3.main import Web3
 
 from ocean_lib.agreements.service_types import ServiceTypes, ServiceTypesNames
+from ocean_lib.services.consumer_parameters import ConsumerParameters
 
 logger = logging.getLogger(__name__)
 
@@ -301,79 +300,3 @@ class Service:
         ] = trusted_algo_publishers
         self.compute_values["allowNetworkAccess"] = allow_network_access
         self.compute_values["allowRawAlgorithm"] = allow_raw_algorithm
-
-
-class ConsumerParameters:
-    def __init__(
-        self,
-        name: str,
-        type: str,
-        label: str,
-        required: bool,
-        default: str,
-        description: str,
-        options: Optional[List[str]] = None,
-    ) -> None:
-
-        fn_args = locals().copy()
-        for attr_name in ConsumerParameters.required_attrs():
-            setattr(self, attr_name, fn_args[attr_name])
-
-        if options is not None and not isinstance(options, list):
-            raise TypeError("Options should be a list")
-
-        self.options = options
-
-    @classmethod
-    def from_dict(
-        cls, consumer_parameters_dict: Dict[str, Any]
-    ) -> "ConsumerParameters":
-        """Create a ConsumerParameters object from a JSON string."""
-        cpd = copy.deepcopy(consumer_parameters_dict)
-        missing_attributes = [
-            x for x in ConsumerParameters.required_attrs() if x not in cpd.keys()
-        ]
-
-        if missing_attributes:
-            raise TypeError(
-                "ConsumerParameters is missing the keys "
-                + ", ".join(missing_attributes)
-            )
-
-        required = cpd["required"] if "required" in cpd else None
-
-        return cls(
-            cpd["name"],
-            cpd["type"],
-            cpd["label"],
-            bool(strtobool(required)) if isinstance(required, str) else required,
-            cpd["default"],
-            cpd["description"],
-            cpd.pop("options", None),
-        )
-
-    @enforce_types
-    def as_dictionary(self) -> Dict[str, Any]:
-        """Return the consume parameters object as a python dictionary."""
-
-        result = {
-            attr_name: getattr(self, attr_name)
-            for attr_name in ConsumerParameters.required_attrs()
-        }
-
-        if self.options is not None:
-            result["options"] = self.options
-
-        return result
-
-    @staticmethod
-    @enforce_types
-    def required_attrs():
-        return [
-            "name",
-            "type",
-            "label",
-            "required",
-            "default",
-            "description",
-        ]
