+++ ocean_lib/services/test/test_service.py
@@ -5,7 +5,7 @@
 import pytest
 
 from ocean_lib.assets.asset import Asset
-from ocean_lib.services.service import ConsumerParameters, Service
+from ocean_lib.services.service import Service
 from tests.resources.ddo_helpers import (
     get_sample_algorithm_ddo,
     get_sample_ddo,
@@ -111,41 +111,6 @@ def test_additional_information():
             "nested_dict": {"some_key": "value"},
         },
     }
-
-
-@pytest.mark.unit
-def test_consumer_parameters():
-    """Tests the Consumer Parameters key/object."""
-    cp_dict = {
-        "name": "test_key",
-        "type": "string",
-        "label": "test_key_label",
-        "required": True,
-        "default": "value",
-        "description": "this is a test key",
-        "options": ["a", "b"],
-    }
-
-    cp_object = ConsumerParameters.from_dict(cp_dict)
-    assert cp_object.as_dictionary() == cp_dict
-
-    cp_dict.pop("options")
-    cp_object = ConsumerParameters.from_dict(cp_dict)
-    assert cp_object.as_dictionary() == cp_dict
-
-    cp_dict["required"] = "false"  # explicitly false, not missing
-    cp_object = ConsumerParameters.from_dict(cp_dict)
-    assert cp_object.as_dictionary()["required"] is False
-
-    cp_dict["options"] = "not an array"
-    with pytest.raises(TypeError):
-        cp_object = ConsumerParameters.from_dict(cp_dict)
-
-    cp_dict.pop("options")
-    cp_dict.pop("type")
-    cp_dict.pop("label")
-    with pytest.raises(TypeError, match="is missing the keys type, label"):
-        cp_object = ConsumerParameters.from_dict(cp_dict)
 
 
 @pytest.mark.unit
